+++ src/internal/range-generator.ts
